Added preliminary env support.
Also changed to requiring to be run *outside* of a node so that we can
do parallel execution of matrix env stuff, and required that this only
be run from a CpsScmFlowDefinition so that we have access to "scm" for
checkout.

Tests work fine...when you run "mvn test
-Dtest=TravisPipelineConverterDSL". They explode horribly when run
with just "mvn test", with infinite loops of git stuff happening. I
have no idea why, but will figure that out eventually.